+++ src/styles/customCSS.ts
@@ -9,13 +9,21 @@ export const fittedStyle: React.CSSProperties = {
   margin: 0,
 }
 
+export const overflowWrapStyle: React.CSSProperties = {
+  overflow: 'overlay',
+  overflowWrap: 'break-word',
+}
+
+export const primaryActiveStyle: React.CSSProperties = {
+  backgroundColor: '#464775',
+}
+
+export const secondaryActiveStyle: React.CSSProperties = {
+  backgroundColor: 'rgba(0,0,0,.25)',
+}
+
 export const truncateStyle: React.CSSProperties = {
   overflow: 'hidden',
   textOverflow: 'ellipsis',
   whiteSpace: 'nowrap',
-}
-
-export const overflowWrapStyle: React.CSSProperties = {
-  overflow: 'overlay',
-  overflowWrap: 'break-word',
 }